Compute and print cost basis when writing output.
When making a sale, it's important for tax purposes to track the cost basis.
In theory MSSB would do this for us, but in practice they do not seem to.
The 2017 1099-DIV attributed the entire sale value as cost basis and nearly
gave me a heart attack.

Cost basis is accumulated piecewise, to avoid rounding the total until the
complete basis value is known.